jwt 使用

+++ admin/src/main/java/com/igeek/admin/service/base/UserTokenService.java
@@ -2,7 +2,7 @@ package com.igeek.admin.service.base;
 
 public interface UserTokenService {
 
-    String createSaveToken(Integer userId);
+    String createSaveToken(Long userId);
 
     boolean logout(Integer userId);
 }
